+++ src/components/common/saves.js
@@ -1,0 +1,8 @@
+export const SaveObjects = {
+    settings: {
+        close: 'settings-close'
+    },
+    banners: {
+        performance: 'performanceBanner'
+    }
+};