+++ src/data/mocks/disclaimer.js
@@ -1,0 +1,9 @@
+const mockDisclaimerData = `This is a website of companies offering financial services - SFoxInsurance, SFoxInsurance Liising.
+                  
+Before entering into any agreement read and conditions of respective service. Consult a specialist, where necessary.
+SFoxInsurance does not provide a credit advisory service for the purposes of the Insurance Intermediates Act.
+The borrower makes the decision of taking out a loan, who accesses the suitability of the loan product and contractual terms to his/her personal
+loan interest, need and financial situation on the basis of the information and warnings presented by the bank
+and is responsible for the consequences related to concluding the agreement.`;
+
+export default mockDisclaimerData;