+++ src/data/mocks/performance.js
@@ -1,0 +1,8 @@
+const performanceBannerData = `With 700 000 private customers and more than 60 000 corporate 
+and organizational customers. This makes us Sweden's largest bank 
+in terms of numbers and gives us a leading position in our other home
+markets of Estonia, US, Latvia, Lithuania. As a major bank, we are a significant part
+of the financial system and play an important role in the local communities we serve.
+We are dedicated to helping our customers, our shareholders and society as whole stay financially sound and sustainable.`;
+
+export default performanceBannerData;